+++ src/kernel_evm/evm_execution/src/handler.rs
@@ -18,6 +18,7 @@ use crate::EthereumError;
 use crate::PrecompileSet;
 use alloc::borrow::Cow;
 use alloc::rc::Rc;
+use core::cmp::min;
 use core::convert::Infallible;
 use evm::executor::stack::Log;
 use evm::gasometer::{GasCost, Gasometer, MemoryCost};
@@ -1581,21 +1582,34 @@ impl<'a, Host: Runtime> Handler for EvmHandler<'a, Host> {
         is_static: bool,
         context: Context,
     ) -> Capture<CallOutcome, Self::CallInterrupt> {
-        if let Err(err) = self.record_cost(target_gas.unwrap_or(0)) {
+        let gas_limit = target_gas.map(|gas| {
+            // Part of EIP-150: https://github.com/ethereum/EIPs/blob/master/EIPS/eip-150.md
+            let max_gas_limit = if self.config.call_l64_after_gas {
+                gas - gas / 64
+            } else {
+                gas
+            };
+
+            min(max_gas_limit, self.gas_remaining())
+        });
+
+        if let Err(err) = self.record_cost(gas_limit.unwrap_or(0)) {
             log!(
                 self.host,
                 Debug,
                 "Not enought gas for call. Required at least: {:?}",
-                target_gas
+                gas_limit
             );
 
             return Capture::Exit((
-                ExitReason::Fatal(ExitFatal::CallErrorAsFatal(ExitError::OutOfGas)),
+                ExitReason::Fatal(ExitFatal::Other(Cow::from(
+                    "Out of gas before recursive call",
+                ))),
                 vec![],
             ));
         }
 
-        if let Err(err) = self.begin_inter_transaction(is_static, target_gas) {
+        if let Err(err) = self.begin_inter_transaction(is_static, gas_limit) {
             return Capture::Exit((ethereum_error_to_exit_reason(err), vec![]));
         }