+++ src/kernel_evm/CHANGES.md
@@ -25,6 +25,7 @@
 
 - The internal kernel version is now automatically updated at each upgrade. (!10321)
 - Genesis parent hash is 0xff..ff, all parent hashes are now unique. (!10332)
+- Fix the bug that nested contract calls returns `OutOfGas` (!10283)
 
 ### Breaking changes